+++ src/realm/object-store/impl/object_notifier.hpp
@@ -22,22 +22,28 @@
 #include <realm/object-store/impl/collection_notifier.hpp>
 
 #include <realm/keys.hpp>
+#include <realm/table.hpp>
 
 namespace realm {
 
 namespace _impl {
 class ObjectNotifier : public CollectionNotifier {
 public:
-    ObjectNotifier(std::shared_ptr<Realm> realm, TableKey table, ObjKey obj);
+    ObjectNotifier(std::shared_ptr<Realm> realm, TableKey table_key, ObjKey obj_key);
 
 private:
-    TableKey m_table;
-    ObjKey m_obj;
+    TableRef m_table;
+    TableKey m_table_key;
+    ObjKey m_obj_key;
     TransactionChangeInfo* m_info;
 
-    void run() override;
+    void run() override REQUIRES(!m_callback_mutex);
+
+    void do_attach_to(Transaction& sg) override;
 
     bool do_add_required_change_info(TransactionChangeInfo& info) override;
+
+    bool object_was_deleted(const ObjectChangeSet& object_change_set);
 };
 } // namespace _impl
 } // namespace realm