+++ src/realm/sync/client.cpp
@@ -82,7 +82,6 @@ public:
 
 private:
     _impl::ClientFileAccessCache m_file_access_cache;
-    const std::shared_ptr<ChangesetCooker> m_default_changeset_cooker;
     const bool m_one_connection_per_session;
     util::network::Trigger m_actualize_and_finalize;
     util::network::DeadlineTimer m_keep_running_timer;
@@ -422,8 +421,6 @@ private:
     std::int_fast64_t m_staged_upload_mark = 0, m_staged_download_mark = 0;
     std::int_fast64_t m_reached_upload_mark = 0, m_reached_download_mark = 0;
 
-    static auto choose_cooker(ClientImpl& client, const Session::Config& config) -> std::shared_ptr<ChangesetCooker>;
-
     static ClientImplBase::Session::Config make_session_impl_config(SyncTransactReporter&, Session::Config&);
 
     void do_initiate(ProtocolEnvelope, std::string server_address, port_type server_port,
@@ -523,7 +520,6 @@ inline SessionWrapperQueue::~SessionWrapperQueue()
 inline ClientImpl::ClientImpl(Client::Config config)
     : ClientImplBase{make_client_impl_base_config(config)}                            // Throws
     , m_file_access_cache{config.max_open_files, config.disable_sync_to_disk, logger} // Throws
-    , m_default_changeset_cooker{std::move(config.changeset_cooker)}
     , m_one_connection_per_session{config.one_connection_per_session}
     , m_keep_running_timer{get_service()} // Throws
 {
@@ -1120,8 +1116,7 @@ void SessionImpl::on_resumed()
 
 SessionWrapper::SessionWrapper(ClientImpl& client, std::string realm_path, Session::Config config)
     : m_client{client}
-    , m_file_slot{client.m_file_access_cache, std::move(realm_path), config.encryption_key,
-                  choose_cooker(client, config)}
+    , m_file_slot{client.m_file_access_cache, std::move(realm_path), config.encryption_key}
     , m_protocol_envelope{config.protocol_envelope}
     , m_server_address{std::move(config.server_address)}
     , m_server_port{config.server_port}
@@ -1504,16 +1499,6 @@ inline void SessionWrapper::report_sync_transact(VersionID old_version, VersionI
     if (m_sync_transact_handler)
         m_sync_transact_handler(old_version, new_version); // Throws
 }
-
-
-auto SessionWrapper::choose_cooker(ClientImpl& client, const Session::Config& config)
-    -> std::shared_ptr<ChangesetCooker>
-{
-    if (config.changeset_cooker)
-        return config.changeset_cooker;
-    return client.m_default_changeset_cooker;
-}
-
 
 auto SessionWrapper::make_session_impl_config(SyncTransactReporter& transact_reporter, Session::Config& config)
     -> ClientImplBase::Session::Config