+++ test/object-store/sync/app.cpp
@@ -279,7 +279,7 @@ TEST_CASE("app: UsernamePasswordProviderClient integration", "[sync][app]") {
         processed = false;
         auto user = app->current_user();
         REQUIRE(user);
-        CHECK(user->user_profile().email == email);
+        CHECK(user->user_profile().email() == email);
     }
 
     SECTION("confirm user") {
@@ -388,7 +388,7 @@ TEST_CASE("app: UsernamePasswordProviderClient integration", "[sync][app]") {
         processed = false;
         auto user = app->current_user();
         REQUIRE(user);
-        CHECK(user->user_profile().email == email);
+        CHECK(user->user_profile().email() == email);
 
         CHECK(user->state() == SyncUser::State::LoggedIn);
 
@@ -409,7 +409,7 @@ TEST_CASE("app: UsernamePasswordProviderClient integration", "[sync][app]") {
         CHECK(user->state() == SyncUser::State::Removed);
         CHECK(app->current_user() != user);
         user = app->current_user();
-        CHECK(user->user_profile().email == email);
+        CHECK(user->user_profile().email() == email);
         CHECK(user->state() == SyncUser::State::LoggedIn);
 
         app->remove_user(user, [&](Optional<app::AppError> error) {
@@ -2287,14 +2287,14 @@ TEST_CASE("app: sync integration", "[sync][app]") {
         auto pred = [](const SyncError& error) {
             return error.error_code.category() == util::websocket::websocket_close_status_category();
         };
-        // If we haven't gotten an error in more than 2 minutes, then something has gone wrong
+        // If we haven't gotten an error in more than 5 minutes, then something has gone wrong
         // and we should fail the test.
         REQUIRE_NOTHROW(timed_wait_for(
             [&] {
                 std::lock_guard<std::mutex> lk(sync_error_mutex);
                 return std::any_of(sync_errors.begin(), sync_errors.end(), pred);
             },
-            std::chrono::minutes(2)));
+            std::chrono::minutes(5)));
 
         auto captured_error = [&] {
             std::lock_guard<std::mutex> lk(sync_error_mutex);
@@ -2306,6 +2306,61 @@ TEST_CASE("app: sync integration", "[sync][app]") {
         REQUIRE(captured_error.error_code.category() == util::websocket::websocket_close_status_category());
         REQUIRE(captured_error.error_code.value() == 1009);
         REQUIRE(captured_error.message == "read limited at 16777217 bytes");
+    }
+}
+
+TEST_CASE("app: custom user data integration tests", "[sync][app]") {
+    std::unique_ptr<GenericNetworkTransport> (*factory)() = [] {
+        return std::unique_ptr<GenericNetworkTransport>(new IntTestTransport);
+    };
+    std::string base_url = get_base_url();
+    const std::string valid_pk_name = "_id";
+    REQUIRE(!base_url.empty());
+    auto app_session = create_app(default_app_config(base_url));
+
+    auto app_config = App::Config{app_session.client_app_id,
+                                  factory,
+                                  base_url,
+                                  util::none,
+                                  Optional<std::string>("A Local App Version"),
+                                  util::none,
+                                  "Object Store Platform Tests",
+                                  "Object Store Platform Version Blah",
+                                  "An sdk version"};
+
+    SECTION("custom user data happy path") {
+        TestSyncManager sync_manager(app_config);
+        auto app = sync_manager.app();
+
+        bool processed = false;
+
+        std::shared_ptr<SyncUser> the_user;
+        app->log_in_with_credentials(AppCredentials::anonymous(),
+                                     [&](std::shared_ptr<SyncUser> user, Optional<app::AppError> error) {
+                                         CHECK(user);
+                                         CHECK(!error);
+                                         processed = true;
+                                         the_user = user;
+                                     });
+
+        CHECK(processed);
+        processed = false;
+
+        app->call_function(the_user, "updateUserData", {bson::BsonDocument({{"favorite_color", "green"}})},
+                           [&](auto error, auto response) {
+                               CHECK(error == none);
+                               CHECK(response);
+                               CHECK(*response == true);
+                               processed = true;
+                           });
+        CHECK(processed);
+        processed = false;
+        app->refresh_custom_data(the_user, [&](auto) {
+            processed = true;
+        });
+        CHECK(processed);
+        auto data = *the_user->custom_data();
+        CHECK(data["favorite_color"] == "green");
     }
 }
 
@@ -2686,34 +2741,64 @@ TEST_CASE("app: login_with_credentials unit_tests", "[sync][app]") {
                                   "Object Store Platform Tests",
                                   "Object Store Platform Version Blah",
                                   "An sdk version"};
+        auto base_path = util::make_temp_dir() + app_name;
+        {
+            auto conf = TestSyncManager::Config(config);
+            conf.should_teardown_test_directory = false;
+            conf.base_path = base_path;
+            TestSyncManager tsm(conf, {});
+            auto app = tsm.app();
 
-        TestSyncManager tsm(TestSyncManager::Config(config), {});
-        auto app = tsm.app();
+            app->log_in_with_credentials(
+                realm::app::AppCredentials::anonymous(),
+                [&](std::shared_ptr<realm::SyncUser> user, util::Optional<app::AppError> error) {
+                    CHECK(user);
+                    CHECK(!error);
 
-        app->log_in_with_credentials(realm::app::AppCredentials::anonymous(),
-                                     [&](std::shared_ptr<realm::SyncUser> user, util::Optional<app::AppError> error) {
-                                         CHECK(user);
-                                         CHECK(!error);
+                    CHECK(user->identities().size() == 2);
+                    CHECK(user->identities()[0].id == UnitTestTransport::identity_0_id);
+                    CHECK(user->identities()[1].id == UnitTestTransport::identity_1_id);
+                    SyncUserProfile user_profile = user->user_profile();
 
-                                         CHECK(user->identities().size() == 2);
-                                         CHECK(user->identities()[0].id == UnitTestTransport::identity_0_id);
-                                         CHECK(user->identities()[1].id == UnitTestTransport::identity_1_id);
-                                         SyncUserProfile user_profile = user->user_profile();
+                    CHECK(user_profile.name() == profile_0_name);
+                    CHECK(user_profile.first_name() == profile_0_first_name);
+                    CHECK(user_profile.last_name() == profile_0_last_name);
+                    CHECK(user_profile.email() == profile_0_email);
+                    CHECK(user_profile.picture_url() == profile_0_picture_url);
+                    CHECK(user_profile.gender() == profile_0_gender);
+                    CHECK(user_profile.birthday() == profile_0_birthday);
+                    CHECK(user_profile.min_age() == profile_0_min_age);
+                    CHECK(user_profile.max_age() == profile_0_max_age);
 
-                                         CHECK(user_profile.name == profile_0_name);
-                                         CHECK(user_profile.first_name == profile_0_first_name);
-                                         CHECK(user_profile.last_name == profile_0_last_name);
-                                         CHECK(user_profile.email == profile_0_email);
-                                         CHECK(user_profile.picture_url == profile_0_picture_url);
-                                         CHECK(user_profile.gender == profile_0_gender);
-                                         CHECK(user_profile.birthday == profile_0_birthday);
-                                         CHECK(user_profile.min_age == profile_0_min_age);
-                                         CHECK(user_profile.max_age == profile_0_max_age);
+                    processed = true;
+                });
 
-                                         processed = true;
-                                     });
+            CHECK(processed);
+        }
+        App::clear_cached_apps();
+        // assert everything is stored properly between runs
+        {
+            auto conf = TestSyncManager::Config(config);
+            conf.base_path = base_path;
+            TestSyncManager tsm(conf, {});
+            auto app = tsm.app();
+            REQUIRE(app->all_users().size() == 1);
+            auto user = app->all_users()[0];
+            CHECK(user->identities().size() == 2);
+            CHECK(user->identities()[0].id == UnitTestTransport::identity_0_id);
+            CHECK(user->identities()[1].id == UnitTestTransport::identity_1_id);
+            SyncUserProfile user_profile = user->user_profile();
 
-        CHECK(processed);
+            CHECK(user_profile.name() == profile_0_name);
+            CHECK(user_profile.first_name() == profile_0_first_name);
+            CHECK(user_profile.last_name() == profile_0_last_name);
+            CHECK(user_profile.email() == profile_0_email);
+            CHECK(user_profile.picture_url() == profile_0_picture_url);
+            CHECK(user_profile.gender() == profile_0_gender);
+            CHECK(user_profile.birthday() == profile_0_birthday);
+            CHECK(user_profile.min_age() == profile_0_min_age);
+            CHECK(user_profile.max_age() == profile_0_max_age);
+        }
     }
 
     SECTION("login_anonymous bad") {