+++ CHANGELOG.md
@@ -18,6 +18,27 @@
 * Remove the STATE message request and handling from the client and test sever. The current server was always responding with empty state anyhow. Clients will now consider receiving this message as an error because they never request it.
 * Handling for async open, involving state files is mostly removed.
 * When a client reset occured, we would create a metadata Realm to handle state download but this isn't needed anymore so it is removed.
+* Removed the unused ChangesetCooker from the sync client. ([#4811](https://github.com/realm/realm-core/pull/4811))
+
+----------------------------------------------
+
+# 11.1.0 Release notes
+
+### Enhancements
+* Change notifications can now be filtered via a key path. This keypath is passed via the `add_notification_callback` on `Object`, `Set`, `List` and `Result`.
+  If such a key path was provided when adding a notification callback it will only ever be executed when a changed property was covered by this filter.
+  Multiple key path filters can be provided when adding a single notification callback.
+
+### Fixed
+* User profile now correctly persists between runs.
+ 
+### Breaking changes
+* User profile fields are now methods instead of fields.
+
+-----------
+
+### Internals
+* Added UniqueFunction type ([#4815](https://github.com/realm/realm-core/pull/4815))
 
 ----------------------------------------------